device bug

+++ train.py
@@ -150,7 +150,7 @@ def weighted_train(epoch, train_loader, p_net, p_optimizer, g_net, g_optimizer, 
         
         if(method == 'weighted_loss_xy'):
             oh = data.numpy().repeat(class_dim, axis=0)
-            one_hot = torch.cat([torch.from_numpy(oh), torch.from_numpy(one_hot)], dim=1).float()
+            one_hot = torch.cat([torch.from_numpy(oh).float(), torch.from_numpy(one_hot).float()], dim=1).float()
         else:
             one_hot = torch.from_numpy(one_hot).float()
         target_concat = torch.from_numpy(target.numpy().repeat(class_dim, axis=0)).float()
